Update - Delete Grades

+++ app/Http/Controllers/Grades/GradeController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Grades;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreGrades;
+use App\Models\Classroom;
 use App\Models\Grade;
 use Illuminate\Http\Request;
 
@@ -29,13 +30,7 @@ class GradeController extends Controller
     public function store(StoreGrades $request)
     {
 
-        if (Grade::where('Name->ar', $request->Name)->orWhere('Name->en', $request->Name_en)->exists()) {
-
-            return redirect()->back()->withErrors(trans('Grades_trans.exists'));
-        }
-
         try {
-
             $validated = $request->validated();
             $Grade = new Grade();
             /*
@@ -87,10 +82,19 @@ class GradeController extends Controller
      */
     public function destroy(Request $request)
     {
+        $MyClass_id = Classroom::where('Grade_id', $request->id)->pluck('Grade_id');
 
-        $Grades = Grade::findOrFail($request->id)->delete();
-        toastr()->error(trans('messages.Delete'));
-        return redirect()->route('Grades.index');
+        if ($MyClass_id->count() == 0) {
+
+            $Grades = Grade::findOrFail($request->id)->delete();
+            toastr()->error(trans('messages.Delete'));
+            return redirect()->route('Grades.index');
+        } else {
+
+            toastr()->error(trans('Grades_trans.delete_Grade_Error'));
+            return redirect()->route('Grades.index');
+
+        }
 
     }
